refactored the docling-core for the document.py
