+++ docling_core/types/doc/common/__init__.py
@@ -1,0 +1,1 @@
+"""Shared, cross-cutting value models for the Docling document model."""
