+++ docling_core/types/doc/items/__init__.py
@@ -1,0 +1,1 @@
+"""Document-tree node item types for the Docling document model."""
