+++ docling_core/types/doc/items/picture/__init__.py
@@ -1,0 +1,1 @@
+"""Picture item and its data models."""
