+++ docling_core/types/doc/items/picture/charts.py
@@ -1,0 +1,169 @@
+"""Picture chart annotation models."""
+
+from typing import Literal
+
+from pydantic import BaseModel
+
+from docling_core.types.doc.common.annotations import BaseAnnotation
+from docling_core.types.doc.items.table.table_data import TableData
+
+
+class ChartLine(BaseModel):
+    """Represents a line in a line chart.
+
+    Attributes:
+        label (str): The label for the line.
+        values (list[tuple[float, float]]): A list of (x, y) coordinate pairs
+            representing the line's data points.
+    """
+
+    label: str
+    values: list[tuple[float, float]]
+
+
+class ChartBar(BaseModel):
+    """Represents a bar in a bar chart.
+
+    Attributes:
+        label (str): The label for the bar.
+        values (float): The value associated with the bar.
+    """
+
+    label: str
+    values: float
+
+
+class ChartStackedBar(BaseModel):
+    """Represents a stacked bar in a stacked bar chart.
+
+    Attributes:
+        label (list[str]): The labels for the stacked bars. Multiple values are stored
+            in cases where the chart is "double stacked," meaning bars are stacked both
+            horizontally and vertically.
+        values (list[tuple[str, int]]): A list of values representing different segments
+            of the stacked bar along with their label.
+    """
+
+    label: list[str]
+    values: list[tuple[str, int]]
+
+
+class ChartSlice(BaseModel):
+    """Represents a slice in a pie chart.
+
+    Attributes:
+        label (str): The label for the slice.
+        value (float): The value represented by the slice.
+    """
+
+    label: str
+    value: float
+
+
+class ChartPoint(BaseModel):
+    """Represents a point in a scatter chart.
+
+    Attributes:
+        value (Tuple[float, float]): A (x, y) coordinate pair representing a point in a
+            chart.
+    """
+
+    value: tuple[float, float]
+
+
+class PictureChartData(BaseAnnotation):
+    """Base class for picture chart data.
+
+    Attributes:
+        title (str): The title of the chart.
+    """
+
+    title: str
+
+
+class PictureLineChartData(PictureChartData):
+    """Represents data of a line chart.
+
+    Attributes:
+        kind (Literal["line_chart_data"]): The type of the chart.
+        x_axis_label (str): The label for the x-axis.
+        y_axis_label (str): The label for the y-axis.
+        lines (list[ChartLine]): A list of lines in the chart.
+    """
+
+    kind: Literal["line_chart_data"] = "line_chart_data"
+    x_axis_label: str
+    y_axis_label: str
+    lines: list[ChartLine]
+
+
+class PictureBarChartData(PictureChartData):
+    """Represents data of a bar chart.
+
+    Attributes:
+        kind (Literal["bar_chart_data"]): The type of the chart.
+        x_axis_label (str): The label for the x-axis.
+        y_axis_label (str): The label for the y-axis.
+        bars (list[ChartBar]): A list of bars in the chart.
+    """
+
+    kind: Literal["bar_chart_data"] = "bar_chart_data"
+    x_axis_label: str
+    y_axis_label: str
+    bars: list[ChartBar]
+
+
+class PictureStackedBarChartData(PictureChartData):
+    """Represents data of a stacked bar chart.
+
+    Attributes:
+        kind (Literal["stacked_bar_chart_data"]): The type of the chart.
+        x_axis_label (str): The label for the x-axis.
+        y_axis_label (str): The label for the y-axis.
+        stacked_bars (list[ChartStackedBar]): A list of stacked bars in the chart.
+    """
+
+    kind: Literal["stacked_bar_chart_data"] = "stacked_bar_chart_data"
+    x_axis_label: str
+    y_axis_label: str
+    stacked_bars: list[ChartStackedBar]
+
+
+class PicturePieChartData(PictureChartData):
+    """Represents data of a pie chart.
+
+    Attributes:
+        kind (Literal["pie_chart_data"]): The type of the chart.
+        slices (list[ChartSlice]): A list of slices in the pie chart.
+    """
+
+    kind: Literal["pie_chart_data"] = "pie_chart_data"
+    slices: list[ChartSlice]
+
+
+class PictureScatterChartData(PictureChartData):
+    """Represents data of a scatter chart.
+
+    Attributes:
+        kind (Literal["scatter_chart_data"]): The type of the chart.
+        x_axis_label (str): The label for the x-axis.
+        y_axis_label (str): The label for the y-axis.
+        points (list[ChartPoint]): A list of points in the scatter chart.
+    """
+
+    kind: Literal["scatter_chart_data"] = "scatter_chart_data"
+    x_axis_label: str
+    y_axis_label: str
+    points: list[ChartPoint]
+
+
+class PictureTabularChartData(PictureChartData):
+    """Base class for picture chart data.
+
+    Attributes:
+        title (str): The title of the chart.
+        chart_data (TableData): Chart data in the table format.
+    """
+
+    kind: Literal["tabular_chart_data"] = "tabular_chart_data"
+    chart_data: TableData
