+++ docling_core/types/doc/items/table/__init__.py
@@ -1,0 +1,1 @@
+"""Table item and its data models."""
